requirements.txt
cassandra-driver
python-dotenv
pandas
